+++ src/util.h
@@ -231,7 +231,7 @@ ERROR_CODE util_blockAlloc(void**, const uint_fast64_t);
 
 ERROR_CODE util_unMap(void*, const uint_fast64_t);
 
-void util_concatenate(char*, const size_t, const char*, const size_t, const char*, const size_t);
+void util_concatenate(char*, const uint_fast64_t, const char*, const uint_fast64_t, const char*, const uint_fast64_t);
 
 uint_fast16_t util_byteArrayTo_uint16(const int8_t*);
